feat: support repos that are not part of an organization
Why: prior to this, this module wouldn't work with repos that are part
of a personal GitHub account, only those that are part of an
organization. This makes it hard to test integrating with this module

This is a bit of a hack, since it uses the terraform `count` mechanism
nonintuitively to only load the org data if the owner_id isn't provided.

Alternative considered: just skip the org lookup altogether and require
the caller to provide the owner ID. This sounds good to me, but we have
to migrate anyone who's using this module.