Output GAR repo ID and location (#14)
This allows callers to easily build IAM bindings for the GAR repo.

Responding to anticipated objections:

Q: the GAR repo ID and name are inputs, so the user should already know
their values.
A: Not if they're using the default variable values.

Q: Isn't this information already available in the
full/name/with/slashes?
A: Parsing it would be gross.

Q: Does the caller really need to know this?
A: Yes. Cloud Run Service Agents need read access to the GAR repo.