Remove oauth2_client_secret_sha256 attribute (#15)
I attempted to leverage the
[iap_config](https://github.com/abcxyz/terraform-modules/blob/main/modules/gclb_cloud_run_backend/variables.tf#L51)
to associate IAP with a load balancer. The `oauth2_client_secret_sha256`
field was set to required but the [iap client
resource](https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/iap_client)
does not export this attribute. I tried to use the terraform sha256
method on the actual clientId but that also does not work (I didnt think
it would but tried it anyway).

Without this change this is the error received 
```
Can't configure a value for "iap.0.oauth2_client_secret_sha256": its value will be decided automatically based on the result of applying this configuration.
```

By making this value optional and only passing the `oauth2_client_id`
and `oauth2_client_secret`, it becomes valid. This indicates to me that
IAP instance will associate to the load balancer. If im missing
something obvious, please let me know!