Require google provider >= 6.0 (#75)
Follow-up to https://github.com/abcxyz/terraform-modules/pull/59. That
change updated the `gclb_cloud_run_backend` module to set a new field,
which I believe was needed due to a change in the
[`google_compute_backend_service`](https://registry.terraform.io/providers/hashicorp/google/6.0.0/docs/resources/compute_backend_service#enabled)
module in version 6.0. If a `gclb_cloud_run_backend` gets an update to
one of those modules but not the other, however, Terraform plans will
fail, as I described in
https://github.com/abcxyz/terraform-modules/pull/59#issuecomment-2438963484.

This change will force a version bump for the
`google_compute_backend_service` provider to ensure that these modules
stay in sync.